requirements.txt
Jinja2==2.7.1
PyYAML==3.10
argparse==1.2.1
nosier==1.1
pypandoc==0.7.0
